move glossiness to PBR function to visual
